Wip - Handle makefile non-mangle-aware nature
The makefile is non-mingw-mangle aware (as far
as I can tell). So when it creates 'git-instaweb'
from git-instaweb.sh the values it uses for the
location of the perl interpreter and the document
root aren't absolute / valid for when the script
runs.  If we can change this somehow then this change
could be removed, and a change to the makefile made
instead.  Unfortunately I'm butting up against the
limits of my knowledge here, and could do with a hand :(